+++ CHANGELOG.rst
@@ -2,6 +2,10 @@ Changelog
 =========
 
 
+- Modify behavior of ``default`` argument to ``Cache.get()`` so that if ``default`` is a callable and the cache key is missing, then it will be called and its return value will be used as the value for cache key and subsequently be set as the value for the key in the cache. (**breaking change**)
+- Add ``default`` argument to ``Cache()`` that can be used to override the value for ``default`` in ``Cache.get()``.
+
+
 v0.9.0 (2018-03-31)
 -------------------